Added some extra trees and background elements to the overall scene, making a immersive arena for the player and the boyds to fly in.
Changed the materials of the boyds body, tails wings etc for personal preference.

Modified the sensitivity of the pitch, yaw and roll of the boyds to help it make it sharper turns and cuts.

Added a stop button so that the player could magically stop midair letting the boyds pass so that they can still get behind them and shoot them.

Increased the speed for the spped up allowing the player to catch up to the boyds.